Migrated files to split Meteor format
