+++ .meteor/packages
@@ -25,3 +25,4 @@ check@1.2.3
 standard-minifier-css@1.2.0_1
 standard-minifier-js@1.2.0_1
 shell-server
+ecmascript
