+++ imports/api/scores.js
@@ -1,0 +1,3 @@
+import { Meteor } from 'meteor/meteor';
+
+export const Scores = new Mongo.Collection('scores');
