+++ server/fireball.js
@@ -1,0 +1,6 @@
+import { Meteor } from 'meteor/meteor';
+import '../imports/api/scores.js';
+
+Meteor.startup(function () {
+  // code to run on server at startup
+});
